bam/writer/record: Write smallest value for integer data field
The BAM format defines more sizes to encode integer data field values.
When writing BAM record data from SAM record data, the smallest value
type is chosen to encode an integer.

This matches the observed behavior of samtools 1.11 when converting the
following data:

```
ZV:i:-2147483648  ZN:Z:i32::MIN
ZV:i:-2147483647  ZN:Z:i32::MIN + 1
ZV:i:-32769       ZN:Z:i16::MIN - 1
ZV:i:-32768       ZN:Z:i16::MIN
ZV:i:-32767       ZN:Z:i16::MIN + 1
ZV:i:-129         ZN:Z:i8::MIN - 1
ZV:i:-128         ZN:Z:i8::MIN
ZV:i:-127         ZN:Z:i8::MIN + 1
ZV:i:-1           ZN:Z:-1
ZV:i:0            ZN:Z:0
ZV:i:1            ZN:Z:1
ZV:i:126          ZN:Z:i8::MAX - 1
ZV:i:127          ZN:Z:i8::MAX
ZV:i:128          ZN:Z:i8::MAX + 1
ZV:i:254          ZN:Z:u8::MAX - 1
ZV:i:255          ZN:Z:u8::MAX
ZV:i:256          ZN:Z:u8::MAX + 1
ZV:i:32766        ZN:Z:i16::MAX - 1
ZV:i:32767        ZN:Z:i16::MAX
ZV:i:32768        ZN:Z:i16::MAX + 1
ZV:i:65534        ZN:Z:u16::MAX - 1
ZV:i:65535        ZN:Z:u16::MAX
ZV:i:65536        ZN:Z:u16::MAX + 1
ZV:i:2147483646   ZN:Z:i32::MAX - 1
ZV:i:2147483647   ZN:Z:i32::MAX
```

+++ noodles-bam/src/writer.rs
@@ -449,7 +449,7 @@ mod tests {
 
         assert_eq!(
             fields.next().transpose()?,
-            Some(Field::new(SamTag::AlignmentHitCount, Value::Int32(1)))
+            Some(Field::new(SamTag::AlignmentHitCount, Value::UInt8(1)))
         );
 
         assert!(fields.next().is_none());
